requirements.txt
asgiref==3.11.1
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
crispy-tailwind==1.0.3
cssselect2==0.9.0
Django==5.2.11
django-crispy-forms==2.5
django-htmx==1.27.0
djangorestframework==3.16.1
fonttools==4.61.1
idna==3.11
pillow==12.1.1
pycparser==3.0
pydyf==0.12.1
pyphen==0.17.2
python-decouple==3.8
reportlab==4.4.10
requests==2.32.5
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.3
urllib3==2.6.3
weasyprint==68.1
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.1
